# docs/swagger.yaml
components:
  schemas:
...
    github_com_array_banking-api_internal_dto.LoginRequest:
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
...
    github_com_array_banking-api_internal_dto.RegisterRequest:
      properties:
        email:
          type: string
        firstName:
          maxLength: 100
          minLength: 1
          type: string
        lastName:
          maxLength: 100
          minLength: 1
          type: string
        password:
          minLength: 12
          type: string
      required:
      - email
      - firstName
      - lastName
      - password
      type: object
...
    github_com_array_banking-api_internal_dto.TokenResponse:
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
      type: object
...
    github_com_array_banking-api_internal_errors.ErrorDetail:
      properties:
        code:
          type: string
        details:
          items:
            type: string
          type: array
          uniqueItems: false
        message:
          type: string
        trace_id:
          type: string
      type: object
    github_com_array_banking-api_internal_errors.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorDetail'
      type: object
...
    message:
      properties:
        message:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Production-quality banking REST API for developer assessment and interviewing.
    Provides core banking functionality including identity management, account operations,
    customer management, and transaction processing.
  title: Array Banking API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /auth/login:
    post:
      description: Authenticate user with email and password, receive JWT access and
        refresh tokens
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.LoginRequest'
        description: Login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.TokenResponse'
          description: Login successful with JWT tokens
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Validation error - AUTH_001
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Invalid credentials - AUTH_002
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Account locked - AUTH_006
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: System error - SYSTEM_001 or SYSTEM_002
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Invalidate user's access token and refresh token. Requires Bearer
        token in Authorization header.
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/message'
                properties:
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    type: object
                type: object
          description: Logout successful
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Unauthorized - AUTH_004 or AUTH_005
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: System error - SYSTEM_001
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Authentication
  /auth/refresh:
    post:
      description: Get a new access token and refresh token pair using a valid refresh
        token
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refresh_token:
                  type: string
              type: object
        description: Refresh token
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.TokenResponse'
          description: Token refreshed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Validation error - AUTH_001
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Invalid refresh token - AUTH_003
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: System error - SYSTEM_001 or SYSTEM_002
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      description: Create a new user account with email, password, and personal information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.RegisterRequest'
        description: Registration details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/message'
                properties:
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    type: object
                type: object
          description: User created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Validation error - AUTH_001 (Invalid request body)
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: Customer already exists - CUSTOMER_001
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: System error - SYSTEM_001 or SYSTEM_002
      summary: Register a new user
      tags:
      - Authentication
...